requirements.txt
google-auth-oauthlib
requests
yt-dlp # Requires srt_fix plugin, install via: pip install https://github.com/bindestriche/srt_fix/archive/refs/heads/master.zip
python-dotenv
google-cloud-speech
google-cloud-storage
google-generativeai
google-api-python-client
google-auth-httplib2
dotenv
webvtt-py